test/lpddr4: move run_simulation wrapper to phy_common.py
